feat: Add JWT authentication and authorization
This commit adds JWT authentication and authorization to the blog application. The REST_FRAMEWORK settings are updated to include the 'rest_framework_simplejwt.authentication.JWTAuthentication' class as the default authentication class. Additionally, the 'rest_framework.permissions.IsAuthenticated' class is set as the default permission class. The SIMPLE_JWT settings are also configured to specify the access token and refresh token lifetimes, as well as other JWT-related configurations.